Close all delegates during OnFilterRemoved().
Should resolve all crashes unless we're getting into a state where
AddDelegate/RemoveDelegate is called by a delegate for which Start()
had not yet been called.

I suspect that's unlikely since at the time of RemoveFilter() the
RenderThread should be in its dying throes, leaving no one else alive
to issue such a call.

I've also removed the CHECK() entirely since it's not thread safe.
The message filters are not destructed on the io thread.

This is all a product of AudioOutputIPC not being reference counted
while IPC::ChannelProxy::MessageFilter is.  We end up with devices
which hold uncounted references.

BUG=178499
TEST=none

Review URL: https://codereview.chromium.org/12378035

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@185644 0039d316-1c4b-4281-b951-d872f2087c98