Separate invalidator and sync client ID (part 1/2)
This change implements support for setting the sync client ID and
invalidator client ID separately.  The two IDs are managed separately
and both of them are sent up to the server.  This change includes some
additional changes to support the transition, such as a new field in
the about:sync UI.

At this point, the IDs are not different.  Both are initialized from the
same source: the sync database.  The work to store and manage the
invalidator's ID separate from sync was begun in r180907 and will be
finished in a separate commit.  We will not be able to complete that
work until the server supports separate IDs.

BUG=124142


Review URL: https://chromiumcodereview.appspot.com/12256033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@185721 0039d316-1c4b-4281-b951-d872f2087c98