Remove CompositorThread dependency on WebThread
content:: creates and manages the compositor thread itself, it doesn't need to
use a WebThread type or use WebCompositorSupport to manage threaded-ness. This
cuts the dependency of content::CompositorThread to just depend on base::Thread.
The message loop is bound to in content::RenderWidgetCompositor.


Review URL: https://chromiumcodereview.appspot.com/12386078

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@185945 0039d316-1c4b-4281-b951-d872f2087c98