Omnibox: In HQP, Keep All URL and Title Parsing Consistent
HistoryQuick provider formats/unescapes URLs when building its index.
This makes matching better.  However, when running the scoring algorithm
it doesn't format URLs in the same way.  This means that the matches
the scoring algorithm finds may not be in comparable positions to the
word_starts information stored in the index.

This fixes it.

It's a bit icky because we have to pass a |languages| variable everywhere
because it controls how URLs were interpreted at the time of indexing.

Regarding testing: because HistoryQuick provider doesn't currently use word starts in scoring, it's a bit of a hassle to write a test for this at the moment.  In a follow-up changelist, where I make HistoryQuick provider use word boundaries, I'll make sure there's an end-to-end test of storing a URL (and its word starts) and verifying that the matching algorithm is finding word boundaries in the expected places.

BUG=177818


Review URL: https://chromiumcodereview.appspot.com/12321098

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@185978 0039d316-1c4b-4281-b951-d872f2087c98