cc: Make DelegatedRendererLayerImpl::SetRenderPasses() private.
This method was now public only for tests. SetFrameData() should be
used instead. Implemented a SetFrameDataForRenderPasses() method on
the FakeDelegatedRendererLayerImpl class. This class constructs a
frame data out of a set of render passes, making a bogus set up
TransferableResources for each resource in the render passes, and
gives the whole frame to the real DelegatedRendererLayerImpl.

Tested by all the delegated renderer layer tests.

R=piman
BUG=123444


Review URL: https://chromiumcodereview.appspot.com/12388082

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@185997 0039d316-1c4b-4281-b951-d872f2087c98