Reinstate files from a previous changelist that were reverted expirementally.
These files were LGTMed and checked in, but reverted speculatively to fix a
build problem.  The problem was not fixed by the revert, so the rest of the CL
got checked back in, but somehow these files got missed.

These files are test data for a disabled browser test (so they cannot affect
the current build or browser test in theory).


BUG=179442


Review URL: https://chromiumcodereview.appspot.com/12379050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@186085 0039d316-1c4b-4281-b951-d872f2087c98