Fix detection of a volume mode in the action choice dialog in Files.app.
Sometimes, the icon was incorrect because of the not fully initialized volume manager. Initialization of the volume manager is asynchronous, so before using it directly, we should wait. This patch adds a ready event to notify listeners about the fact that the mount points have been initialized.

NOTRY=True
TEST=Enter an usb drive and an sd card with DCIM directory. A sd icon should appear in the dialog.
BUG=178350


Review URL: https://chromiumcodereview.appspot.com/12387077

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@186604 0039d316-1c4b-4281-b951-d872f2087c98