Fix memory leak in BrowserAccessibilityManager::UpdateNode.
This memory leak was only happening when a fatal error was detected,
because the method was returning false in the middle of updating the
tree, sometimes leaving an allocated object without a parent.

Now the method only has two returns: once at the top, if a duplicate
child id is found, before any new memory has been allocated, and again
at the very bottom, if a reparented child id was found and skipped.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/12570004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@186776 0039d316-1c4b-4281-b951-d872f2087c98