Attaching a DOM Activity Logger to all extension scripts
Adds a mechanism for attaching (using the WebKit API) a DOMActivityLogger
to each world that is executing extension scripts. (This would include the
main world in the case of an extension background page.) A logger is only 
attached if extension activity logging is enabled and if there does not 
already exist a logger attached to the world. Each logger object has a log
method that marshalls it's arguments and sends them over to the extension
activity log on the browser side. A HashMap in WebCore owns the logger
objects. Currently the loggers are not invoked, a later WebCore patch will
appropriately configure the DOM bindings for each world so that they invoke
the corresponding logger object on every access.

BUG=160987, 160989

TBR=brettw@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12390076

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@186845 0039d316-1c4b-4281-b951-d872f2087c98