Adjust DCHECK() for being current in GLSurface::GetExtensions().
With virtual contexts we avoid (potentially expensive) MakeCurrent()
calls when our shared context is already current and we are switching
to an offscreen surface.
Since GLContextVirtual::IsCurrent() already contains all that logic
to pretend a context is current, change the DCHECK() accordingly.
It was not handling all cases in the current form.

BUG=180575


Review URL: https://chromiumcodereview.appspot.com/12621005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@187061 0039d316-1c4b-4281-b951-d872f2087c98