Remove BluetoothAdapterFactory::Create
We never used the ability to pull up a particular Bluetooth Adapter
by address or device name, and it was never really supported in
Windows - and is equally difficult in Mac and in Chromium OS when
using the Bluetooth Smart-capable backend.

So drop it.

BUG=none
TEST=device_unittests


Review URL: https://chromiumcodereview.appspot.com/12413011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@187476 0039d316-1c4b-4281-b951-d872f2087c98