The proper review is in https://codereview.chromium.org/12310101/.
r187677 was reverted due to it broke the chromeos bots. And it seems some chromeos bots do not have the libpulse-dev headers files. Instead of fixing the bots, I added back the USE_PULSEAUDIO macro to disable pulse for chromeos.

This patch enables pulse as the default IO handling in Chrome.

Right now we have the dynamic linking for pulse library, for the machine without pulse, it will automatically fall back to ALSA.

In order to enable pulse as the default IO, libpulse-dev package is required to build Chrome on linux. Previously libpulse-dev is only required if building chromeOS, right now it is also required by building chrome.

Note, we still have some small issues for the pulse implementations, so this patch needs to be landed after https://codereview.chromium.org/12328097/,
https://codereview.chromium.org/12316131/, https://codereview.chromium.org/12310102/ and
https://codereview.chromium.org/12613005/

TBR=dalecurtis@chromium.org
BUG=178101
TEST=compile on linux

Review URL: https://codereview.chromium.org/12646006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@187834 0039d316-1c4b-4281-b951-d872f2087c98