Remove ant stuff dealing with library dependencies
This is mostly dealing with the code for library dependencies. Our
library dependencies are handled through gyp, and not by ant. The
dependency task that is removed calculates only empty paths for each of
its arguments (some are already set above and are immutable).


BUG=158821
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12595008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@187911 0039d316-1c4b-4281-b951-d872f2087c98