Remote RDP sessions, rather than the console, if curtain-mode is configured.
This CL adds a new DesktopSession implementation that starts an RDP terminal and attaches to it. Attaching to the terminal launches the DesktopSessionAgent into the corresponding RDP session, and a new agent will be launched each time the terminal's active session is switched, e.g. if the user logs in to an existing session, in the same way as when remoting the host's physical console.

The terminal will be configured to match client dimensions as reported by the Chromoting host.

BUG=137696

Review URL: https://chromiumcodereview.appspot.com/12544020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188283 0039d316-1c4b-4281-b951-d872f2087c98