When Chrome enters the fullscreen mode: 1) If the panel is created with initial focus, as in the case that the extension icon is clicked, the panel will be shown. 2) If the panel is created without initial focus, as in the case of new incoming message, the panel will not be shown.
When Chrome leaves the fullscreen mode, all panels that are delayed for showing will be shown now.

This patch also fixed the initial fullscreen detection problem for OSX. If Chrome enters fullscreen mode before any panel is created, we need to find out if chrome is in fullscreen or not.

BUG=180795
TEST=Manual test by creating panels when Chrome is in fullscreen mode


Review URL: https://chromiumcodereview.appspot.com/12430013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188300 0039d316-1c4b-4281-b951-d872f2087c98